requirements.txt
requests==2.31.0
python-telegram-bot==21.6
pymongo==3.12.0
python-dotenv==0.21.1
telethon==1.31.1
